[Web Animations] Expose the id property on Animation https://bugs.webkit.org/show_bug.cgi?id=181450 <rdar://problem/36383600>
Reviewed by Dean Jackson.

Source/WebCore:

Expose the "id" property on Animation and handle the "id" property on the optional KeyframeAnimationOptions object
passed to Element.animate(). All of the WPT tests related to this property are now passing.

* animation/WebAnimation.h:
* animation/WebAnimation.idl:
* dom/Element.cpp:
(WebCore::Element::animate):

LayoutTests:

Mark all tests related to the "id" property as passing.

* http/wpt/web-animations/interfaces/Animatable/animate-expected.txt:
* http/wpt/web-animations/interfaces/Animation/id-expected.txt:
* http/wpt/web-animations/interfaces/Animation/idlharness-expected.txt:

Identifier: 197329@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@226697 268f45cc-cd09-0410-ab3c-d52691b4dbfc