REGRESSION (r226818): API test WebKitLegacy.LoadInvalidURLRequest is failing https://bugs.webkit.org/show_bug.cgi?id=181595
This test was testing behavior of an invalid URL that WebCore parses but NSURL does not.
It was using example.com<> but after r226479 < and > are considered invalid by WebCore.
I change the < and > to $ to make this test pass and continue to test what it used to.

* TestWebKitAPI/Tests/mac/LoadInvalidURLRequest.html:
* TestWebKitAPI/Tests/mac/LoadInvalidURLRequest.mm:
(-[LoadInvalidURLWebFrameLoadDelegate webView:didFailProvisionalLoadWithError:forFrame:]):



Identifier: 197456@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@226897 268f45cc-cd09-0410-ab3c-d52691b4dbfc