JSC should use a speculation fence on VM entry/exit https://bugs.webkit.org/show_bug.cgi?id=181991
Reviewed by JF Bastien and Mark Lam.

Source/JavaScriptCore:

This adds a WTF::speculationFence on VM entry and exit.

For a microbenchmark that just calls a native function (supplied via an Objective-C block) in a
tight loop from JS is a 0% regression on x86 and a 11% regression on ARM64.

* runtime/JSLock.cpp:
(JSC::JSLock::didAcquireLock):
(JSC::JSLock::willReleaseLock):

Source/WTF:

Implement speculationFence as lfence on x86 and isb on ARM64. I'm not sure if isb is
appropriate for all ARM64's.

* wtf/Atomics.h:
(WTF::speculationFence):
(WTF::x86_lfence):



Identifier: 197824@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@227424 268f45cc-cd09-0410-ab3c-d52691b4dbfc